more button changed to have multiple options and added a separate controller
