+++ app/javascript/controllers/more_action_controller.js
@@ -1,0 +1,9 @@
+import { Controller } from "@hotwired/stimulus";
+
+export default class extends Controller {
+  static targets = ["deactivateMenu"];
+
+  toggleDropdown() {
+    this.deactivateMenuTarget.classList.toggle("hidden");
+  }
+}
