+++ app/views/shared/components/_more_action.html.erb
@@ -1,14 +1,22 @@
-<div class="relative mt-1">
-  <label class="flex items-center cursor-pointer rounded-lg bg-line-colour" data-action="click->team-user-status#toggleDropdown">
+<div class="relative mt-1" data-controller="more-action" >
+  <label class="flex items-center cursor-pointer rounded-lg bg-line-colour" data-action="click->more-action#toggleDropdown">
     <span class="icon icon-more-vertical icon-small"></span>
   </label>
-  <div class="absolute bg-white rounded box-shadow-medium hidden mt-3 -ml-20 z-10" data-team-user-status-target="deactivateMenu">
-    <button
-      class="text-xs cursor-pointer px-4 py-2 hover:bg-gray-100"
-      data-action="click->team-user-status#showConfirmation"
-      data-user-status-confirmation-target="deactivateMenu"
-      data-activate="false">
-      Deactivate
-    </button>
+  <div class="absolute bg-white rounded box-shadow-medium hidden mt-3 -ml-20 z-10" data-more-action-target="deactivateMenu" data-user-status-confirmation-target="moreMenu">
+    <% actions.each do |action| %>
+      <% if action == :deactivate %>
+        <button
+          class="text-xs cursor-pointer px-4 py-2 hover:bg-gray-100 w-full text-left"
+          data-action="click->user-status-confirmation#showConfirmation"
+          data-user-status-confirmation-target="deactivateMenu"
+          data-activate="false">
+          Deactivate
+        </button>
+      <% elsif action == :edit %>
+        <button class="text-xs cursor-pointer px-4 py-2 hover:bg-gray-100 w-full text-left">Edit</button>
+      <% elsif action == :delete %>
+        <button class="text-xs cursor-pointer px-4 py-2 hover:bg-gray-100 w-full text-left">Delete</button>
+      <% end %>
+    <% end %>
   </div>
 </div>
