+++ app/views/users/_team_user_info.html.erb
@@ -33,7 +33,7 @@
         <span class="bg-cover inline-block icon-red-circle icon-tiny"></span>
         <p class="text-xs text-danger">Deactivated</p>
       </div>
-      <%= render 'shared/components/more_action' %>
+      <%= render 'shared/components/more_action', actions: %i[deactivate] %>
     </div>
   </div>
   <%= render 'shared/components/confirmation_modal' %>
